+++ templates/src/test/expected-java/tst2/x/ev/EventB.java
@@ -26,6 +26,7 @@ import org.fuin.ddd4j.core.EventType;
 import org.fuin.ddd4j.jsonb.AbstractDomainEvent;
 import org.fuin.objects4j.common.Contract;
 import org.fuin.objects4j.core.KeyValue;
+import org.fuin.objects4j.core.KeyValueEL;
 import org.fuin.objects4j.ui.Examples;
 import p.x.ev.CustomerId;
 
@@ -71,8 +72,8 @@ public final class EventB extends AbstractDomainEvent<CustomerId> {
 
     @Override
     public String toString() {
-        return Objects.requireNonNull(KeyValue.replace("Event B: ${a} [${#entityIdPath}]",
-            new KeyValue("#entityIdPath", getEntityIdPath())
+        return Objects.requireNonNull(KeyValueEL.replace("Event B: ${a} [${entityIdPath}]",
+            new KeyValue("entityIdPath", getEntityIdPath())
             , new KeyValue("a", a)
         ));
     }