+++ templates/src/test/expected-java/tst2/x/except/ExceptionE.java
@@ -22,6 +22,7 @@ import java.util.Objects;
 import org.fuin.objects4j.common.Contract;
 import org.fuin.objects4j.common.UniquelyNumberedException;
 import org.fuin.objects4j.core.KeyValue;
+import org.fuin.objects4j.core.KeyValueEL;
 
 /**
  * Exception E - With CID - Single variable
@@ -39,7 +40,7 @@ public final class ExceptionE extends UniquelyNumberedException {
      * @param a A.
      */
     public ExceptionE(final String a) {
-        super(124, Objects.requireNonNull(KeyValue.replace("Exception C: ${a}",  new KeyValue("a", a))));
+        super(124, Objects.requireNonNull(KeyValueEL.replace("Exception C: ${a}",  new KeyValue("a", a))));
         Contract.requireArgNotNull("a", a);
         
         this.a = a;